pyproject.toml
[tool.poetry]
name = "gidd-server"
version = "1.0"
description = "GIDD server"
authors = ["dev <dev.togglecorp.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.10"
django-environ = "^0.8.1"
psycopg2 = "^2.9.3"
django-cors-headers = "==3.7.0"
strawberry-graphql-django = "^0.2.5"
strawberry-graphql = {extras = ["debug-server"], version = "^0.104.1"}
djangorestframework = "^3.13.1"
django-filter = "^21.1"
drf-nested-routers = "^0.93.4"
factory-boy = "^3.2.1"
pytest-asyncio = "^0.18.3"
django-storages = "^1.12.3"
django-tinymce = "^3.4.0"
django-reversion = "^5.0.0"
gunicorn = "^20.1.0"
boto3 = "^1.22.13"

[tool.poetry.dev-dependencies]
pytest-profiling = "*"
pytest-ordering = "*"
pytest-django = "*"
pytest-icdiff = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
